+++ tests/unit/providers/factories/test_factory_py2_py3.py
@@ -1,9 +1,10 @@
 """Factory provider tests."""
 
+import decimal
 import sys
 
 from dependency_injector import providers, errors
-from pytest import raises
+from pytest import raises, mark
 
 from .common import Example
 
@@ -27,6 +28,20 @@ def test_init_optional_provides():
 def test_set_provides_returns_():
     provider = providers.Factory()
     assert provider.set_provides(object) is provider
+
+
+@mark.parametrize(
+    "str_name,cls",
+    [
+        ("dependency_injector.providers.Factory", providers.Factory),
+        ("decimal.Decimal", decimal.Decimal),
+        ("list", list),
+        (".common.Example", Example),
+        ("test_is_provider", test_is_provider),
+    ],
+)
+def test_set_provides_string_imports(str_name, cls):
+    assert providers.Factory(str_name).provides is cls
 
 
 def test_init_with_valid_provided_type():